Combine 2 imports into one
